Fixes #1: Adding README.md file
This change will fix issue #1.  I am adding a somewhat basic README.md
file to the repo.  This file is my very first time using markdown, it
is the first branch I have created through GitHub, and it will be my
first real commit of a change.

As mentioned in the README file, while I am intending to run all of my
projects through GitHub, the first project is getting it setup, and it
seems fitting that I should be running that through here as well.