Create first webpage and Jekyll directory
Addresses several issues, including:

* #2: Basic HTML homepage.
  * Not including styling, this is just the basic HTML.
* #10: Added Default Jekyll template.
  * **default.html** added to the **_layouts/** directory.
* #18: Create nav menu dynamically
  * Using the **_data/** directory and a **CSV** file to populate the
creation of the menus.  The code itself is contained in the
**header.html** file and is a simple loop that creates an Unordered
List.
* #19: Fixed HTML Issues.
  * So I was not fully understanding how Jekyll bundles everything, and
I thought that each of the fragments that have an **.html** extension
should be a complete document.  Not the case at all actually, and now I
am correctly building them as reusable fragments that will ultimately
be pieced together to create one complete and valid HTML.